frameworks/base/telephony: Fix for correct PIN2/PUK2 menus.
1) Android UI displays PUK2 verification window even when PUK2 is permanently blocked.
2) Android UI displays Change PIN2 window even when PUK2 is permanently blocked.

This is because Android framework is not interpreting PIN2 Status. Hence UI
does not have necessary information to detect whether PUK2 is permanently blocked
or PIN2 is blocked. Added support in framework to interpret PIN2 status and
propagate this information to UI. UI now has all the information and displays
PIN2 and PUK2 verification menus correctly.

CRs-Fixed: 217760,217761