Fix data type of RIL_UNSOL_RESPONSE_NEW_BROADCAST_SMS in RIL.
Fixing incorrect data type of RIL_UNSOL_RESPONSE_NEW_BROADCAST_SMS
in Android RIL.

The Android RIL specifies that the unsolicited response signal
RIL_UNSOL_RESPONSE_NEW_BROADCAST_SMS should contain a raw byte
array (ril.h). However, in the implementation of the Java RIL and
the C++ RIL daemon, the data is unpacked as a string. This causes
data delivered from the network stack to be garbled.

Changing so that both the C++ RIL deamon in platform/hardware/ril
and the Java RIL in frameworks/base follows the specification.

Change-Id: I9b9b26d16cd39422c43d9fa7aadfcff3e60c78ae